+++ packages/terra-application-utility/src/terra-dev-site/doc/example/ApplicationHeaderUtilityExample.jsx
@@ -26,7 +26,7 @@ class ApplicationHeaderUtilityExample extends React.Component {
     return (
       <div style={{ paddingLeft: '4px' }}>
         <div style={{
-          height: '60px', position: 'relative', width: '150px', backgroundColor: '#2481ca',
+          height: '60px', position: 'relative', width: '150px', backgroundColor: '#0079BE',
         }}
         >
           <ApplicationHeaderUtility