+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -3,6 +3,12 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
 
 3.1.0 - (February 5, 2019)
 ------------------