+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -7,10 +7,13 @@ Unreleased
 * EmbeddedContentProviderTestTemplate for testing embedded applications
 
 ### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
 * Updated existing provider examples to use the EmbeddedContentProviderTestTemplate
 * Changed Provider examples to use `.provider` extension to fix provider examples in terra-ui
 
 ### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
 * Dead code tests / examples that tested XFC implementation
 
 3.1.0 - (February 5, 2019)